refactor(functions)!: Make sure only named exports are used (`usernameToColor`)
