+++ src/components/NcAvatar/NcAvatar.vue
@@ -252,8 +252,8 @@ import NcUserStatusIcon from '../NcUserStatusIcon/index.js'
 
 import { getRoute } from '../../components/NcRichText/autolink.ts'
 import { useIsDarkTheme } from '../../composables/index.ts'
-import usernameToColor from '../../functions/usernameToColor/index.js'
 import { getEnabledContactsMenuActions } from '../../functions/contactsMenu/index.ts'
+import { usernameToColor } from '../../functions/usernameToColor/index.ts'
 import { userStatus } from '../../mixins/index.js'
 import { getAvatarUrl } from '../../utils/getAvatarUrl.ts'
 import { getUserStatusText } from '../../utils/UserStatus.ts'
