+++ src/functions/index.ts
@@ -9,5 +9,5 @@ export * from './dialog/index.ts'
 export * from './emoji/index.ts'
 export * from './isDarkTheme/index.ts'
 export * from './preloadImage/index.ts'
-export * from './reference/index.js'
-export { default as usernameToColor } from './usernameToColor/index.js'
+export * from './reference/index.ts'
+export * from './usernameToColor/index.ts'
