+++ styleguide/global.requires.js
@@ -11,7 +11,7 @@ import axios from '@nextcloud/axios'
 import { isA11yActivation } from '../src/functions/a11y/index.ts'
 import { EmojiSkinTone, emojiSearch, emojiAddRecent, getCurrentSkinTone, setCurrentSkinTone } from '../src/functions/emoji/index.ts'
 import { spawnDialog } from '../src/functions/dialog/index.ts'
-import usernameToColor from '../src/functions/usernameToColor/index.ts'
+import { usernameToColor } from '../src/functions/usernameToColor/index.ts'
 import { useIsDarkTheme } from '../src/composables/useIsDarkTheme/index.ts'
 import NcDialog from '../src/components/NcDialog/index.ts'
