+++ tests/unit/functions/usernameToColor/usernameToColor.spec.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 import { describe, expect, it } from 'vitest'
-import usernameToColor from '../../../../src/functions/usernameToColor/index.ts'
+import { usernameToColor } from '../../../../src/functions/usernameToColor/index.ts'
 
 const rgbToHex = (color) => {
 	return '#' + ((1 << 24) + (color.r << 16) + (color.g << 8) + color.b).toString(16).slice(1)
